chore(crypto-js): Move Vodozemac types into their own Rust module.
